Removed old Graph context.
