+++ torch_xla/csrc/ir_util.cpp
@@ -1,0 +1,44 @@
+#include "ir_util.h"
+
+#include "tensorflow/compiler/xla/xla_client/debug_macros.h"
+
+namespace torch_xla {
+namespace ir {
+
+std::vector<Node*> Util::ComputePostOrder(Node* node, EmissionMap* emap) {
+  std::vector<Node*> post_order;
+  std::vector<Node*> queue;
+  queue.push_back(node);
+  while (!queue.empty()) {
+    node = queue.back();
+    auto it = emap->find(node);
+    if (it == emap->end()) {
+      (*emap)[node] = kEmitting;
+
+      for (auto& output : node->operands()) {
+        auto oit = emap->find(output.node);
+        if (oit == emap->end()) {
+          queue.push_back(output.node);
+        } else if (oit->second == kEmitting) {
+          XLA_ERROR() << "Graph loop found at " << *output.node;
+        }
+      }
+    } else if (it->second == kEmitting) {
+      for (auto& output : node->operands()) {
+        auto oit = emap->find(output.node);
+        XLA_CHECK(oit != emap->end() && oit->second == kEmitted)
+            << "Graph loop found at " << *output.node;
+      }
+      emit_status_[node] = kEmitted;
+      post_order.push_back(node);
+      queue.pop_back();
+    } else {
+      XLA_CHECK_EQ(it->second, kEmitted);
+      queue.pop_back();
+    }
+  }
+  return post_order;
+}
+
+}  // namespace ir
+}  // namespace torch_xla
